Refactor server structure: update entry point to core.js, implement cluster mode, and remove deprecated mapper.ts
