+++ server/package.json
@@ -8,11 +8,11 @@
   "license": "MIT",
   "author": "Ankan Saha",
   "type": "commonjs",
-  "main": "lib/core/mapper.js",
+  "main": "lib/core/core.js",
   "scripts": {
     "test": "npm test",
     "build": "tsc",
-    "start": "node lib/core/mapper.js",
+    "start": "node lib/core/core.js",
     "dev": "npm run build && npm start"
   },
   "dependencies": {
